Refactor: break the 3 module couplings blocking a clean forward/inverse split
Prerequisite, behavior-preserving moves toward #105. These break the only
dependency cycles/wrong-direction edges in the package, so the forward kernel
can later be isolated (as extras or a separate repo) without further surgery.

- Merge `utils/data_handling` + `utils/process` into one `tsadar/data/`
  package. They were mutually recursive (data_handling->process and
  process->data_handling); the split was artificial since preprocessing *is*
  data handling. Now the cycle is intra-package.
- Move `postprocess.py` into `tsadar/inverse/`. It is post-*fit* processing
  and imported `inverse.loss_function`/`inverse.loops` while `inverse.fitter`
  imported it back — a true `inverse <-> process` cycle. Now intra-`inverse`.
- Move the interactive runner `calc_vs_data.py` from `forward/` to `inverse/`.
  It imported `inverse.fitter`/`inverse.loss_function`, making `forward` depend
  on `inverse` (wrong direction). `forward` no longer imports `inverse`.

No behavior change: pure file moves + import-path updates (incl. tests).
Verified statically — all intra-tsadar imports resolve, every module compiles,
and forward->inverse / data->inverse / data->forward edges are gone; `core`
kernel remains free of inverse/forward/data/plotting deps.
