+++ tests/test_inverse/test_1d_random.py
@@ -15,7 +15,7 @@ from flatten_dict import flatten, unflatten
 from tsadar.utils import misc
 from tsadar.core.thomson_diagnostic import ThomsonScatteringDiagnostic
 from tsadar.core.modules.ts_params import ThomsonParams, get_filter_spec
-from tsadar.utils.data_handling.calibration import get_scattering_angles
+from tsadar.data.calibration import get_scattering_angles
 
 
 def _perturb_params_(rng, params):
