+++ tests/test_inverse/test_angular_1v.py
@@ -20,7 +20,7 @@ from flatten_dict import flatten, unflatten
 from tsadar.utils import misc
 from tsadar.core.thomson_diagnostic import ThomsonScatteringDiagnostic
 from tsadar.core.modules.ts_params import ThomsonParams, get_filter_spec
-from tsadar.utils.data_handling.calibration import get_scattering_angles, get_calibrations
+from tsadar.data.calibration import get_scattering_angles, get_calibrations
 from tsadar.inverse.loops import label
 
 def _dump_ts_params(td: str, ts_params: ThomsonParams, prefix: str = ""):
