+++ tsadar/__init__.py
@@ -1,3 +1,3 @@
 from .runner import run_for_app
 from .core import ThomsonScatteringDiagnostic, ThomsonParams
-from .utils.data_handling.calibration import get_scattering_angles
+from .data.calibration import get_scattering_angles
