+++ tsadar/data/feature_detector.py
